+++ lib/src/reqif/reqif_spec_objects.dart
@@ -5,7 +5,6 @@
 import 'package:reqif_editor/src/reqif/reqif_attribute_definitions.dart';
 import 'package:reqif_editor/src/reqif/reqif_attribute_values.dart';
 import 'package:reqif_editor/src/reqif/reqif_common.dart';
-import 'package:reqif_editor/src/reqif/reqif_data_types.dart';
 import 'package:reqif_editor/src/reqif/reqif_document.dart';
 import 'package:reqif_editor/src/reqif/reqif_error.dart';
 import 'package:reqif_editor/src/reqif/reqif_spec_object_type.dart';
@@ -16,6 +15,8 @@ class ReqIfSpecificationObject extends ReqIfIdentifiable {
   static const String xmlName = 'SPEC-OBJECT';
   static const String _xmlNameType = 'TYPE';
   static const String _xmlNameSpecObjectTypeRef = 'SPEC-OBJECT-TYPE-REF';
+  static const String _xmlValues = 'VALUES';
+  final ReqIfDocument document;
 
   /// Constructs an instance of this type by parsing an existing ReqIF document.
   /// May throw an error if links cannot be resolved or if the the document
@@ -24,7 +25,7 @@ class ReqIfSpecificationObject extends ReqIfIdentifiable {
   /// [node] must be of type SPEC-OBJECT.
   ///
   /// [document] is a reference to the parent document.
-  ReqIfSpecificationObject.parse(xml.XmlElement element, ReqIfDocument document)
+  ReqIfSpecificationObject.parse(xml.XmlElement element, this.document)
       : super.parse(element, ReqIfElementTypes.specificationObject) {
     _specificationObjectTypeId = getInnerTextOfGrandChildElements(
         node, _xmlNameType, _xmlNameSpecObjectTypeRef);
@@ -36,7 +37,7 @@ class ReqIfSpecificationObject extends ReqIfIdentifiable {
     _specificationObjectType = link;
     buildChildObjects(
         element: element,
-        firstChildTag: 'VALUES',
+        firstChildTag: _xmlValues,
         builder: (inner) {
           switch (inner.name.local) {
             case "ATTRIBUTE-VALUE-STRING":
@@ -99,5 +100,53 @@ class ReqIfSpecificationObject extends ReqIfIdentifiable {
       notFirst = true;
     }
     return buffer.toString();
+  }
+
+  ReqIfAttributeValueXhtml appendXhtmlValue(String ref,
+      [String ns = "reqif-xhtml"]) {
+    xml.XmlBuilder builder = xml.XmlBuilder();
+    builder.element(ReqIfAttributeValueXhtml.xmlName, nest: () {
+      builder.text('\n');
+      builder.element('DEFINITION', nest: () {
+        builder.element('ATTRIBUTE-DEFINITION-XHTML-REF', nest: () {
+          builder.text(ref);
+        });
+        builder.text('\n');
+      });
+      builder.text('\n');
+      builder.element('THE-VALUE', nest: () {
+        builder.element('$ns:div', nest: () {
+          builder.element('$ns:br');
+        });
+      });
+      builder.text('\n');
+    });
+    builder.text('\n');
+    final fragment = builder.buildDocument();
+    final babies = fragment.children;
+
+    final targets = node.findElements(_xmlValues);
+    if (targets.isEmpty) {
+      throw ReqIfError('internal error: $node has no child $_xmlValues');
+    }
+    final toModify = targets.first.children;
+    int lenBefore = toModify.length;
+    for (final baby in babies) {
+      toModify.add(baby.copy());
+    }
+    // add appends at end of list
+    final newNode = toModify[lenBefore];
+    assert(newNode is xml.XmlElement &&
+        newNode.name.local == ReqIfAttributeValueXhtml.xmlName);
+    children.add(ReqIfAttributeValueXhtml.parse(
+        parent: this, element: newNode as xml.XmlElement, document: document));
+    final newValue = children.last;
+    children.sort((a, b) {
+      a as ReqIfAttributeValue;
+      b as ReqIfAttributeValue;
+      return a.column.compareTo(b.column);
+    });
+    updateLastChange();
+    return newValue as ReqIfAttributeValueXhtml;
   }
 }